app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o app/requirements.txt
alembic==1.18.4
    # via uc-catalog-mcp (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via uvicorn
cryptography==46.0.7
    # via
    #   google-auth
    #   pyjwt
databricks-sdk==0.102.0
    # via uc-catalog-mcp (pyproject.toml)
google-auth==2.49.1
    # via databricks-sdk
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   uc-catalog-mcp (pyproject.toml)
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jsonschema==4.26.0
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
mcp==1.27.0
    # via uc-catalog-mcp (pyproject.toml)
numpy==2.4.4
    # via pgvector
pgvector==0.4.2
    # via uc-catalog-mcp (pyproject.toml)
protobuf==6.33.6
    # via databricks-sdk
psycopg==3.3.3
    # via uc-catalog-mcp (pyproject.toml)
psycopg-binary==3.3.3
    # via psycopg
psycopg2-binary==2.9.11
    # via uc-catalog-mcp (pyproject.toml)
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   mcp
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via mcp
pyjwt==2.12.1
    # via mcp
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.24
    # via mcp
pyyaml==6.0.3
    # via uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via databricks-sdk
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
sqlalchemy==2.0.49
    # via
    #   uc-catalog-mcp (pyproject.toml)
    #   alembic
sse-starlette==3.3.4
    # via mcp
starlette==1.0.0
    # via
    #   mcp
    #   sse-starlette
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   mcp
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.44.0
    # via
    #   uc-catalog-mcp (pyproject.toml)
    #   mcp
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
